requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mysql-connector-python==8.3.0
packaging==24.0
psycopg2-binary==2.9.9
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.2
python-benedict==0.33.2
python-decouple==3.8
python-fsutil==0.14.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
